fix cadastrar carona

+++ src/main/java/com/generation/desafio_3_carona/dto/CaronaUpdateDTO.java
@@ -9,7 +9,7 @@ public class CaronaUpdateDTO {
     private LocalDateTime dataHoraPartida;
     private String origem;
     private String destino;
-    private Integer distanciaKm;
+    private Double distanciaKm;
     private int velocidade;
     private int vagas;
     private BigDecimal valorPorPassageiro;
@@ -39,11 +39,11 @@ public class CaronaUpdateDTO {
         this.destino = destino;
     }
 
-    public Integer getDistanciaKm() {
+    public Double getDistanciaKm() {
         return distanciaKm;
     }
 
-    public void setDistanciaKm(Integer distanciaKm) {
+    public void setDistanciaKm(Double distanciaKm) {
         this.distanciaKm = distanciaKm;
     }
